review_visualizer/requirements.txt
Flask>=2.0
google-cloud-bigquery>=3.0
pytest>=7.0
